backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.0.350
langchain-google-genai==0.0.6
google-generativeai==0.3.2
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.5.2 